The Basic Framework for ScimitarX Hydro is ready and is compiling without any error
