+++ src/Integrator/ScimitarX.cpp
@@ -1,14 +1,18 @@
 #include "IO/ParmParse.H"
 #include "Integrator/ScimitarX.H"
-#include "Util/ScimitarX_SetIndex.H"
 #include "BC/BC.H"
 #include "BC/Nothing.H"
 #include "BC/Constant.H"
 #include "IC/SodShock.H"
 #include "Numeric/Stencil.H"
+#include "Util/Util.H"
+#include "Util/ScimitarX_Util.H"
 
 namespace Integrator
 {
+
+// Define the static member variable
+Util::ScimitarX_Util::getVariableIndex ScimitarX::variableIndex;
 
 ScimitarX::ScimitarX(IO::ParmParse& pp) : ScimitarX()
 {
@@ -18,41 +22,39 @@ ScimitarX::ScimitarX(IO::ParmParse& pp) : ScimitarX()
 void
 ScimitarX::Parse(ScimitarX& value, IO::ParmParse& pp)
 {
-    BL_PROFILE("Integrator::ScimitarX::Parse()")
+    BL_PROFILE("Integrator::ScimitarX::Parse()");
     {
+        ScimitarX::SolverTypeManager& setIndex = ScimitarX::SolverTypeManager::getInstance();
 
-    Util::ScimitarX_SetIndex& setIndex = Util::ScimitarX_SetIndex::getInstance();
+        std::string solverTypeStr;
 
-    std::string solverTypeStr;
+        if (pp.query("SolverType", solverTypeStr)) 
+        {
+            auto it = ScimitarX::stringToSolverType.find(solverTypeStr);
+            if (it != ScimitarX::stringToSolverType.end()) {
+                value.solverType = it->second;
 
-    if (pp.query("SolverType", solverTypeStr)) 
-    {
-        auto it = Util::stringToSolverType.find(solverTypeStr);
-        if (it != Util::stringToSolverType.end()) {
-            Util::SolverType solverType = it->second;
+                ScimitarX::variableIndex = setIndex.computeAndAssignVariableIndices(value.solverType);
+                int number_of_components = ScimitarX::variableIndex.NVAR_MAX;
 
-            auto variableIndicesResult = setIndex.computeAndAssignVariableIndices(solverType);
-            int number_of_components = variableIndicesResult.NVAR_MAX;
+                // Call the setupBoundaryConditions function
+                IO::ParmParse bc_pp = setupPVecBoundaryConditions(pp, ScimitarX::variableIndex);
 
-            // Call the setupBoundaryConditions function
-            IO::ParmParse bc_pp = setupPVecBoundaryConditions(pp, variableIndicesResult);
+                // Example: Print consolidated boundary conditions
+                std::string result;
+                if (bc_pp.query("bc.pvec.type.xlo", result)) {
+                    std::cout << "Boundary condition (type.xlo): " << result << std::endl;
+                }
+                if (bc_pp.query("bc.pvec.val.xlo", result)) {
+                    std::cout << "Boundary condition (val.xlo): " << result << std::endl;
+                }
 
-            // Example: Print consolidated boundary conditions
-            std::string result;
-            if (bc_pp.query("bc.pvec.type.xlo", result)) {
-                std::cout << "Boundary condition (type.xlo): " << result << std::endl;
+                value.bc_PVec = new BC::Constant(number_of_components, pp, "bc.pvec");
+                value.bc_Pressure = new BC::Constant(1, pp, "bc.pressure");
+            } else {
+                Util::Abort(__FILE__, __func__, __LINE__, "Invalid SolverType: " + solverTypeStr);
             }
-            if (bc_pp.query("bc.pvec.val.xlo", result)) {
-                std::cout << "Boundary condition (val.xlo): " << result << std::endl;
-            }
-
-            value.bc_PVec = new BC::Constant(number_of_components, pp, "bc.pvec");
-            value.bc_Pressure = new BC::Constant(1, pp, "bc.pressure");
-        } else {
-            Util::Abort(__FILE__, __func__, __LINE__, "Invalid SolverType: " + solverTypeStr);
         }
-    }
-
     }
     
     // Register New Fabs
@@ -75,33 +77,26 @@ ScimitarX::Parse(ScimitarX& value, IO::ParmParse& pp)
     }
     // Initial Conditions
     {
-    std::string type = "constant";
-    pp.query("ic.pvec.type", type);
+        std::string type = "constant";
+        pp.query("ic.pvec.type", type);
 
-    if (type == "sodshock") {
+        if (type == "sodshock") {
+            value.ic_PVec = new IC::SodShock(value.geom, pp, "ic.pvec.sodshock", ScimitarX::variableIndex);
+        } else {
+            Util::Abort(__FILE__, __func__, __LINE__, "Invalid ic.pvec.type: " + type);
+        }
 
-        Util::ScimitarX_SetIndex& setIndex = Util::ScimitarX_SetIndex::getInstance();
-        auto variableIndicesResult = setIndex.computeAndAssignVariableIndices(Util::SolverType::SolveCompressibleEuler);
-       value.ic_PVec = new IC::SodShock(value.geom, pp, "ic.pvec.sodshock", variableIndicesResult);
-
-    } else {
-        Util::Abort(__FILE__, __func__, __LINE__, "Invalid ic.pvec.type: " + type);
-    }
-
-    pp.query("ic.pressure.type", type);
-    if (type == "sodshock") {
-        value.ic_Pressure = new IC::SodShock(value.geom, pp, "ic.pressure.sodshock");
-    } else {
-        Util::Abort(__FILE__, __func__, __LINE__, "Invalid ic.pressure.type: " + type);
-    }
+        pp.query("ic.pressure.type", type);
+        if (type == "sodshock") {
+            value.ic_Pressure = new IC::SodShock(value.geom, pp, "ic.pressure.sodshock");
+        } else {
+            Util::Abort(__FILE__, __func__, __LINE__, "Invalid ic.pressure.type: " + type);
+        }
     } 
 }
 
 void ScimitarX::Initialize(int lev)
 {
-
-//   PVec_mf[lev]->setVal(0.0);     
-//    Pressure_mf[lev]->setVal(0.0);     
     ic_PVec->Initialize(lev, PVec_mf);
     ic_Pressure->Initialize(lev, Pressure_mf);
 
@@ -124,8 +119,8 @@ void ScimitarX::Advance(int lev, Set::Scalar /*time*/, Set::Scalar /*dt*/)
     for (amrex::MFIter mfi(*QVec_mf[lev], amrex::TilingIfNotGPU()); mfi.isValid(); ++mfi) {
         const amrex::Box& bx = mfi.tilebox();
 
-       amrex::Array4<const Set::Scalar> const& QVec_old = (*QVec_old_mf[lev]).array(mfi);
-       amrex::Array4<Set::Scalar> const& QVec     = (*QVec_mf[lev]).array(mfi);
+        amrex::Array4<const Set::Scalar> const& QVec_old = (*QVec_old_mf[lev]).array(mfi);
+        amrex::Array4<Set::Scalar> const& QVec     = (*QVec_mf[lev]).array(mfi);
 
         for (int n = 0; n < number_of_components; ++n) {
             amrex::ParallelFor(bx, [=] AMREX_GPU_DEVICE(int i, int j, int k) {
@@ -159,20 +154,24 @@ void ScimitarX::TagCellsForRefinement(int lev, amrex::TagBoxArray& a_tags, Set::
     amrex::Print() << "Refinement threshold set to: " << refinement_threshold << "\n"; 
 }
 
+void ScimitarX::TimeStepBegin(Set::Scalar a_time, int a_iter) {
+    // Placeholder implementation
+}
 
-    void TimeStepBegin(Set::Scalar a_time, int a_iter){};
-    void TimeStepComplete(Set::Scalar time, int lev) {};
+void ScimitarX::TimeStepComplete(Set::Scalar time, int lev) {
+    // Placeholder implementation
+}
 
-    void Regrid(int lev, Set::Scalar time){};
+void ScimitarX::Regrid(int lev, Set::Scalar time) {
+    // Placeholder implementation
+}
 
-
-
-IO::ParmParse setupPVecBoundaryConditions(IO::ParmParse& pp, Util::ScimitarX_SetIndex::VariableIndicesResult variableIndicesResult)
+IO::ParmParse ScimitarX::setupPVecBoundaryConditions(IO::ParmParse& pp, const Util::ScimitarX_Util::getVariableIndex& variableIndex)
 {
-    int n_components = variableIndicesResult.NVAR_MAX;
+    int n_components = variableIndex.NVAR_MAX;
     std::vector<std::string> component_names(n_components);
 
-    for (const auto& [variable, index] : variableIndicesResult.variableIndexMap) {
+    for (const auto& [variable, index] : variableIndex.variableIndexMap) {
         component_names[index] = std::to_string(variable);
     }
